The_HIVE/Assistent_Imerion/requirements.txt
# requirements.txt
openai>=1.13
pinecone>=3.2
python-dotenv>=1.0
